requirements.txt
discord.py==2.5.2
mariadb==1.1.8
python-dotenv==1.0.0
asyncio-mqtt==0.16.2